Implemented auto miner
This adds a tool that automatically mines blocks whenever a transaction
appears in the mempool. This can be very handy in tests that create
transactions outside of LNP/BP testkit (automated or manual).

It also exposes a handy `mine_blocks` function to mine blocks
explicitly and refactors other code to be implemented in terms of that
function.

Closes #6